% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_partioning.R
\name{abundance_partioning}
\alias{abundance_partioning}
\title{Title}
\usage{
abundance_partioning(N, SE.N, X, groups = NULL, alpha = 0.05,
  print = TRUE)
}
\arguments{
\item{print}{}
}
\description{
Title
}

