% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cades_JuvOutmigrant.R
\name{cades_JuvOutmigrant}
\alias{cades_JuvOutmigrant}
\title{Format juvenile out-migrant DES tables.}
\usage{
cades_JuvOutmigrant(df, detail_meta, smoltEq_meta, alpha = c("0.05", "0.10"))
}
\arguments{
\item{df}{\code{lifestage} dataframe output from \code{get_JuvenileEstimates}.}

\item{detail_meta}{CA DES juvenile outmigrant detail metadata}

\item{smoltEq_meta}{CA DES juvenile outmigrant metadata}

\item{alpha}{Type I error rate. Default is set at 0.05 to produce 95\%
confidence intervals.}
}
\value{
a list with coordinated assessments juvenile outmigrant and juvenile outmigrant detail DES tables
}
\description{
Format juvenile abundance and survival data generated from
  \code{get_JuvenileEstimates} for the coordinated assessments juvenile
  outmigrant and juvenile outmigrant detail DES tables. The function requires
  an odbc connection from behind the NPT firewall to download CAX metadata
  tables from the CDMS database.  Additionally, if a dataframe is not
  supplied a connection to \url{https://cdms.nptfisheries.org} using
  \code{cdmsR::cdmsLogin()} is also required.
}
\examples{
cdmsR::cdmsLogin('your_username', 'your_password')
df <- get_JUVests()
con <- RODBC::odbcConnect(dsn = 'data source', uid = 'your_username', pwd = 'your_password')
cades_JuvOutmigrant(df, odbc_connection = con)
}
