% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cades_NOSA.R
\name{cades_NOSA}
\alias{cades_NOSA}
\title{Format NOSA Coordinated Assessments DES table}
\usage{
cades_NOSA(nosa_data, nosa_meta, IPTDS = FALSE, alpha = c("0.05", "0.10"))
}
\arguments{
\item{nosa_meta}{CA DES NOSA metadata}

\item{IPTDS}{logical for the estimates being from the DABOM output}

\item{alpha}{Type I error rate. Default is set at 0.05 to produce 95\%
confidence intervals.}

\item{df}{dataframe output from \code{get_NOSAests}}
}
\value{

}
\description{
Format NOSA data generated from
  \code{get_NOSAests} for the coordinated assessments NOSA DES table. The function requires a downloaded CAX metadata
  table from NPT servers.
}
