% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_value_added.R
\name{carcass_value_added}
\alias{carcass_value_added}
\title{Title}
\usage{
carcass_value_added(carcass.data, bin_length, recap_mark, target_mark)
}
\arguments{
\item{target_mark}{}
}
\value{

}
\description{
Title
}
