% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_CarcassData.R
\name{clean_carcassData}
\alias{clean_carcassData}
\title{Clean Carcass Data}
\usage{
clean_carcassData(data)
}
\arguments{
\item{data}{raw CDMS carcass dataset from \code{cdmsR::getDatasetView(datastore = 79)} or from CDMS dataset export}
}
\description{
Processes the raw CDMS carcass dataset and adds important fields for summaries (e.g., Survey Year, Reporting Group, Transect Name, Above Weir, Origin, Mark and Recapture etc.)
}
\examples{
car_dat <- cdmsR::getDatasetView(datastore = 79)
clean_carcassData(car_dat)
}
\author{
Ryan N. Kinzer
}
