% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_CarcassData_NEOR.R
\name{clean_carcassData_NEOR}
\alias{clean_carcassData_NEOR}
\title{Clean North East Oregon Carcass Data - from ODFW Access DB}
\usage{
clean_carcassData_NEOR(data)
}
\arguments{
\item{data}{Import text file with:  read.delim(file = 'path_to_file.txt', sep = ',', header = TRUE)}
}
\description{
Processes the raw ODFW Access DB carcass dataset and standardizes to join with clean_carcassData(CDMS_dat)
}
\examples{
clean_carcass_Data(car_dat)

}
\author{
Tyler T. Stright
}
