% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_FallRR.R
\name{clean_FallRR}
\alias{clean_FallRR}
\title{Clean Fall Chinook Run Reconstruction (FCRR) Data}
\usage{
clean_FallRR(data)
}
\arguments{
\item{data}{raw CDMS FCRR dataset from \code{cdmsR::get_FallRR()} or from CDMS dataset export}
}
\description{
Processes raw FCRR data into a more manageable format.
}
\examples{
fcrr_dat <- get_FallRR())
fcrr_clean <- clean_FallRR(fcrr_dat)
}
\author{
Tyler T. Stright
}
