% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_P4data.R
\name{clean_P4data}
\alias{clean_P4data}
\title{Clean P4 Data}
\usage{
clean_P4data(data)
}
\arguments{
\item{data}{raw CDMS RST dataset from get_RSTdata()}
}
\description{
Processes raw P4 data and adds important fields for summaries
}
\examples{
rst_raw <- get_RSTdata()
rst_clean <- clean_RSTdata(rst_raw)
}
\author{
Tyler T. Stright
}
