% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ReddData.R
\name{clean_reddData}
\alias{clean_reddData}
\title{Clean Redd Data}
\usage{
clean_reddData(data)
}
\arguments{
\item{data}{raw CDMS redd dataset from \code{cdmsR::getDatasetView(datastore = 78)} or from CDMS dataset export}
}
\description{
Processes the raw CDMS redd dataset and adds important fields for summaries (e.g., Survey Year, Reporting Group, Transect Name, Above Weir, etc.)
}
\examples{
redd_dat <- cdmsR::getDatasetView(datastore = 78)
clean_reddData(redd_dat)
}
\author{
Ryan N. Kinzer
}
