% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ReddData_NEOR.R
\name{clean_reddData_NEOR}
\alias{clean_reddData_NEOR}
\title{Clean North East Oregon Redd Data - from ODFW Access DB}
\usage{
clean_reddData_NEOR(data)
}
\arguments{
\item{data}{Import text file with:  read.delim(file = 'path_to_file.txt', sep = ',', header = TRUE)}
}
\description{
Processes the raw ODFW Access DB carcass dataset and standardizes to join with clean_reddData(CDMS_dat)
}
\examples{
clean_reddData_NEOR(redd_dat)

}
\author{
Tyler T. Stright
}
