% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_RSTdata.R
\name{clean_RSTdata}
\alias{clean_RSTdata}
\title{Clean Rotary Screw Trapping Data}
\usage{
clean_RSTdata(data)
}
\arguments{
\item{data}{raw CDMS RST dataset from \code{cdmsR::getDatasetView(datastore = 107)} or from CDMS dataset export}
}
\description{
Processes raw P4 data and adds important fields for summaries
}
\examples{
rst_dat <- getDatasetView(datastore = 107)
tmp <- clean_RSTdata(rst_dat)
}
\author{
Tyler T. Stright
}
