% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_WaterTempData.R
\name{clean_WaterTempData}
\alias{clean_WaterTempData}
\title{clean_WaterTempData:}
\usage{
clean_WaterTempData(data)
}
\arguments{
\item{data}{raw CDMS water temperature dataset from \code{cdmsR::get_WaterTempData(start_date = 'mm-dd-yyyy', end_date = 'mm-dd-yyyy')}}
}
\description{
Cleans raw CDMS water temperature data
}
\examples{
wt_raw <- cdmsR::get_WaterTempData()
wt_clean <- clean_WaterTempData(wt_raw)
}
\author{
Tyler T. Stright
}
