% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_WaterTemps.R
\name{clean_WaterTemps}
\alias{clean_WaterTemps}
\title{Clean Water Temperature Data}
\usage{
clean_WaterTemps(data)
}
\arguments{
\item{data}{raw CDMS water temperature dataset from \code{cdmsR::getWaterTemps(start_date = 'mm-dd-yyyy', end_date = 'mm-dd-yyyy')}}
}
\description{
Processes the raw CDMS water temperature dataset
}
\examples{
wt_dat <- cdmsR::getWaterTemps(datastore = ????)
clean_harvestData(harvest_dat)
}
\author{
Tyler T. Stright
}
