% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_WeirData.R
\name{clean_weirData}
\alias{clean_weirData}
\title{Clean Weir Trapping Data}
\usage{
clean_weirData(data)
}
\arguments{
\item{data}{raw CDMS redd dataset from \code{cdmsR::getDatasetView(datastore = 78)} or from CDMS dataset export}
}
\description{
Processes the raw FINS trapping query and adds important fields for summaries (e.g., Survey Year, Reporting Group, Transect Name, Above Weir, Origin, Mark and Recapture etc.)
}
\examples{
con <- RODBC::odbcConnect(dsn = 'Fins', uid = 'guest', pwd = 'guest')
qry <- "SELECT distinct(Facility) FROM FINS_all_trapping"
f <- RODBC::sqlQuery(con, qry) \%>\% mutate_all(as.character) \%>\% pull(Facility)
npt_f <- f[grepl('NPT', f)]
weir_dat <- get_FinsTrapping(npt_f, odbc_connection = con)
tmp <- clean_weirData(weir_dat)
}
\author{
Ryan N. Kinzer
}
