% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_groups.R
\name{cnt_groups}
\alias{cnt_groups}
\title{Count Groups}
\usage{
cnt_groups(.data, .summary_var, ...)
}
\arguments{
\item{.data}{dataset}

\item{.summary_var}{variable to count}

\item{...}{grouping variables}
}
\value{

}
\description{
Count records in a dataset within defined groups.
}
\author{
Ryan N. Kinzer
}
