% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnload_WeirData.R
\name{dnload_WeirData}
\alias{dnload_WeirData}
\title{Download FINS Trapping Data}
\usage{
dnload_WeirData(
  module = c("Trapping", "Release", "Spawning"),
  apikey,
  startDate,
  endDate
)
}
\arguments{
\item{module}{FINS module to query}

\item{apikey}{user defined API key provided by FINS development team}

\item{startDate}{query start date}

\item{endDate}{query end data}
}
\value{

}
\description{
Download FINS Trapping Data
}
