% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emigrant_abundance.R
\name{emigrant_abundance}
\alias{emigrant_abundance}
\title{Estimate juvenile emigrant abundance from M-R data.}
\usage{
emigrant_abundance(
  data,
  alpha = 0.05,
  iter = 1000,
  print = TRUE,
  save_file = FALSE,
  file_name = NULL
)
}
\arguments{
\item{data}{An R dataframe or "example.csv" file containing at least three fields of data; the
count of unMed Cs (\code{C}), Med fish (\code{M}) released
upstream to conduct trap efficiency trials, and Rd (\code{R}) trap
efficiency fish. Field names are required to be; C, M, R.  All other fields are ignored.}

\item{alpha}{Type I error rate.  Default is set at 0.05 to produce 95\%
confidence intervals.}

\item{iter}{The number of bootstrap iterations to be run.  Default is set at
1000.}

\item{print}{Logical value. Default value is TRUE. Should results be printed to the console?}

\item{save_file}{Logical value. Default value is FALSE.  Should the results be saved as a .csv file to the current working directory?}

\item{file_name}{A character string with the desired name of the outputed .csv file.  The saved file name
will also have the month, day and year of creation appended to the file_name.}
}
\value{
A list of point and uncertainty estimates by each strata.
}
\description{
\code{emigrant_abundance} generates stratified juvenile abundance
  estimates from M-R data collected at rotary screw traps (RST).
  The function operates similar to a Gauss based program developed by
  Steinhorst et al. (2004). Point estimates are derived using the Bailey
  estimator and uncertainty is obtained by bootstrapping unMed capatures
  and Med Rs.
}
\section{Warning}{
 Standard error and confidence intervals will not match GAUSS estimates
  perfectly because of bootstrapped sample distribution.
}

\examples{
# A single strata
est1 <- emigrant_abundance(data.frame(C = 76, M = 68, R = 7), alpha = 0.05, iter = 1000, print = TRUE)
names(est1)
est1$Strata

# Multiple strata
unM <- c(76, 128, 82, 61, 350, 74)
M <- c(68, 117, 70, 42, 282, 64)
recap <- c(7, 10, 9, 18, 48, 10)

dat <- data.frame(C = unM, M = M, R = recap)
est_strata <- emigrant_abundance(data = dat, alpha = .05,iter = 1000, print=TRUE)
est_strata$Strata

est_csv <- emigrant_abundance(data = "example.csv", alpha = .05,iter = 1000, print=TRUE)

}
\author{
Ryan N. Kinzer
}
