% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_propagation.R
\name{error_propagation}
\alias{error_propagation}
\title{Title}
\usage{
error_propagation(
  fx = "multiply",
  type = "both_random",
  X,
  SE.X,
  Y,
  SE.Y = NULL,
  alpha = 0.05
)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
Title
}
