% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_abundance.R
\name{est_abundance}
\alias{est_abundance}
\title{Estimate Abundance}
\usage{
est_abundance(
  n1,
  n2,
  m2,
  method = c("adjusted Peterson"),
  CItype = c("normal", "binomial", "Poisson", "hypergeometric"),
  alpha = 0.05
)
}
\arguments{
\item{n1}{a vector of the number of fish marked and released at the first capture occasion; M}

\item{n2}{a vector of the number of fish captured at the second occasion; C}

\item{m2}{a vector of the number of fish captured at the second occasion with a mark; R}

\item{method}{M/R estimate type; "adjusted peterson',}

\item{CItype}{c('normal','binomial', 'poisson', 'hypergeometric')}

\item{alpha}{type I error rate.  Default is set at 0.05 to produce 95\%
confidence intervals.}
}
\description{
Estimate abundance using the adjusted Peterson formula for a closed
 population (Chapman 1951).
}
\examples{
adjPeterson(n1 = c(100, 1000), n2 = c(50, 500), m2 = c(25, 250))
}
\author{
Ryan N. Kinzer
}
