% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_abundance_carcass.R
\name{est_abundance_carcass}
\alias{est_abundance_carcass}
\title{Carcass Expansion Using Recovery Rate}
\usage{
est_abundance_carcass(n, g, SE.g, groups = NULL, alpha = 0.05, print = TRUE)
}
\arguments{
\item{n}{number of carcasses collected}

\item{g}{recovery rate of carcasses calculated from marks released at weir and recaptured carcasses collected}

\item{SE.g}{standard error of recovery rate}

\item{alpha}{type I error rate.  Default is set at 0.05 to produce 95\%
confidence intervals}
}
\description{
Carcass Expansion Using Recovery Rate
}
\examples{
est_abundance_carcass(100, .5, .05)
}
\author{
Ryan N. Kinzer
}
