% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_abundance_fpr.R
\name{est_abundance_fpr}
\alias{est_abundance_fpr}
\title{Fish per Redd Expansion}
\usage{
est_abundance_fpr(R, N, SE_N, R_D = NULL, x = NULL, n = NULL, alpha = 0.05)
}
\arguments{
\item{R}{redds upstream associated with the estimate of fish, or fish located in}

\item{N}{total abundance of the fish group in similar spatial locations as redds}

\item{SE_N}{standard error of abundance estimate}

\item{R_D}{redds downstream}

\item{x}{number of fish downstream that were in the same group as N}

\item{n}{total number of fish downstream that need to be expanded}

\item{alpha}{type I error rate.  Default is set at 0.05 to produce 95\%
confidence intervals}
}
\description{
Calculates fish per redd expansion estimate for fish upstream of a monitoring site. The function then uses the expansion to create an estimate of the same type of fish downstream of the monitoring site. Depending on the group of fish in the expansion calculation, the resulting downstream estimate may not include all groups of fish downstream of the site. To include all groups of fish, set x and n parameters to calculate the proportion of the initial group downstream, and then expand the downstream estimate by the proportion.
}
\examples{
est_abundance_fpr(100, 200, 5, alpha = .05)
}
\author{
Ryan N. Kinzer
}
