% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_abundance_sgs.R
\name{est_abundance_sgs}
\alias{est_abundance_sgs}
\title{Redd Expansion Using Females and Prespawn Mortality}
\usage{
est_abundance_sgs(
  R,
  p_F,
  SE_F,
  p_Psp,
  SE_Psp,
  CI_type = c("delta", "boot"),
  alpha = 0.05,
  iterations = 1000
)
}
\arguments{
\item{R}{total redd number}

\item{p_F}{female proportion}

\item{SE_F}{standard error of female proportion}

\item{p_Psp}{prespawn mortality proportion}

\item{SE_Psp}{standard error of prespawne mortality}

\item{CI_type}{either a delta or bootstrap method for precision estimates}

\item{alpha}{type I error rate.  Default is set at 0.05 to produce 95\%
confidence intervals}

\item{iterations}{bootstrap iterations}
}
\description{
Estimate abundance of fish using the Chasco et al. (2014) redd
 expansion. The calculation expands redds using female and prespawn
  mortality proportions.
}
\examples{
est_abundance_sgs(100, .5, .02, .1, .01, alpha = .05)
}
\author{
Ryan N. Kinzer
}
