% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_emigrants.R
\name{est_emigrants}
\alias{est_emigrants}
\title{Estimate juvenile emigrant abundance from M-R data.}
\usage{
est_emigrants(
  data,
  unmark,
  mark,
  recap,
  alpha = 0.05,
  iter = 1000,
  boot_values = FALSE,
  print = FALSE
)
}
\arguments{
\item{data}{An R dataframe containing at least three fields of data; the count of "unmark",  "mark", and "recap" fish captured in the trap. Each row of the #' data is treated as an independent strata. Grouping variables can also be included in the data (e.g., trap season or life-stage).}

\item{alpha}{Type I error rate.  Default is set at 0.05 to produce 95\%
confidence intervals.}

\item{iter}{The number of bootstrap iterations to be run.  Default is set at
1000.}
}
\value{
A list of point and uncertainty estimates by each strata.
}
\description{
\code{emigrant_abundance} generates stratified juvenile abundance
  estimates from M-R data collected at rotary screw traps (RST).
  The function operates similar to a Gauss based program developed by
  Steinhorst et al. (2004). Point estimates are derived using the Bailey
  estimator and uncertainty is obtained by bootstrapping unmarked captures
  and marked recaptures.
}
\section{Warning}{
 Standard error and confidence intervals will not match GAUSS estimates
  perfectly because of bootstrapped sample distribution.
}

\examples{
u <- c(76, 128, 82, 61, 350, 74)
m <- c(68, 117, 70, 42, 282, 64)
r <- c(7, 10, 9, 18, 48, 10)

dat <- data.frame('u' = u, 'm' = m, 'r' = r)
est_strata <- est_emigrants(data = dat, alpha = .05,iter = 1000, print=TRUE)
est_strata$Strata

}
\author{
Ryan N. Kinzer
}
