% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_group_p_sum.R
\name{est_group_p_sum}
\alias{est_group_p_sum}
\title{Estimate Group Proportions from Counts}
\usage{
est_group_p_sum(.data, .summary_var, .cnt_var, alpha = 0.05, ...)
}
\arguments{
\item{.data}{dataset}

\item{.summary_var}{variable to estimate proportions for}

\item{alpha}{type I error rate.  Default is set at 0.05 to produce 95\%
confidence intervals.}

\item{...}{grouping variables}

\item{.sum}{variable to sum}
}
\value{

}
\description{
Sum a count field and estimate proportions with precision for defined grouping variables.
}
\examples{
car_df <- cdmsR::getDatasetView(datastoreID = 79)
car_dat <- clean_ReddData(car_df)
est_group_p(car_dat, Sex, alpha = 0.05, SurveyYear, StreamName, Species, Run)
}
\author{
Ryan N. Kinzer
}
