% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_portion.R
\name{est_portion}
\alias{est_portion}
\title{Estimate the portion of a total.}
\usage{
est_portion(N, SE_N, phat, SE_phat, alpha = 0.05)
}
\arguments{
\item{N}{total escapement or abundance}

\item{SE_N}{standard error of N}

\item{phat}{vector of probabilities}

\item{SE_phat}{vector of standard errors for probabilities}

\item{alpha}{type I error rate.  Default is set at 0.05 to produce 95%
confidence intervals.}
}
\description{
Portion a total number \(i.e. escapement or abundance\) into its parts using estimated proportions. \code{est_portion} will return the estimated portion and the associated precision of each estimate.  Currently the portions are calculated assuming independence of the proportions.  In the future this function should be calcuated using a multinomial distribtion.
}
\examples{
Nhat = 1000
SE_N = 50
phat = c(.2,.5,.3)
SE_phat = c(.01,.03,.01)
est_portion(N, SE_N, phat, SE_phat)

}
\author{
Ryan N. Kinzer
}
