% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_proportion.R
\name{est_proportion}
\alias{est_proportion}
\title{Estimate a proporiton and associated uncertainty.}
\usage{
est_proportion(x, n, method = c("score", "wald"), alpha = 0.05)
}
\arguments{
\item{x, n}{An integer value of the count of successes (\code{x}) and trials (\code{n}).}

\item{method}{'wald', 'exact', 'score'. The default method is 'score'.}

\item{alpha}{Type I error rate.  Default is set at 0.05 to produce 95%
confidence intervals.}
}
\description{
\code{estimate_proportion} estimates a proporiton using the
maximum likelihood estimator.  Uncertainty is estimated by either the
Wald or Score methods.
}
\examples{
estimate_proportion(x =  10, n = 100, method = 'score', alpha = 0.05)

}
\author{
Ryan N. Kinzer
}
