% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_proportions.R
\name{est_proportions}
\alias{est_proportions}
\title{Estimate Multiple Proportions}
\usage{
est_proportions(x, alpha = 0.05)
}
\arguments{
\item{x}{a vector of quantities for each estimated group}

\item{alpha}{type I error rate.  Default is set at 0.05 to produce 95\%
confidence intervals.}
}
\value{

}
\description{
Estimate multiple proportions with uncertainty assuming a
binomial distribution. In the future this function will likely be changed to
use a multinomial distribution.
}
\examples{
est_proportions(x = c(10,20,70))
}
