% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_proportion.R
\name{estimate_proportion}
\alias{estimate_proportion}
\title{Estimate a proporiton and associated uncertainty.}
\usage{
estimate_proportion(x, n, method = "score", alpha = 0.05, print = TRUE)
}
\arguments{
\item{x, n}{An integer value of the count of successes (\code{x}) and trials (\code{n}).}

\item{method}{'wald', 'exact', 'score'. The default method is 'score'.}

\item{alpha}{Type I error rate.  Default is set at 0.05 to produce 95%
confidence intervals.}

\item{print}{Should results be printed to the console? Default is TRUE.}
}
\value{
A list of point and uncertainty estimates.
}
\description{
\code{estimate_proportion} estimates a proporiton using the
maximum likelihood estimator.  Uncertainty is estimated by either the
exact, Wald or Score methods.
}
\examples{
estimate_proportion(x =  10, n = 100, method = 'score', alpha = 0.05,
print = TRUE)

}
\author{
Ryan N. Kinzer
}

