% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_IPTDS_NosaDES.R
\name{format_IPTDS_NosaDES}
\alias{format_IPTDS_NosaDES}
\title{Format Natural Origin Spawner Abundance (NOSA) DES tables for IPTDS.}
\usage{
format_IPTDS_NosaDES(
  df = NULL,
  alpha = c("0.05", "0.10"),
  odbc_connection,
  cdms_url = "https://npt-cdms.nptfisheries.org"
)
}
\arguments{
\item{df}{\code{excellent} dataframe output from \code{get_NOSAestimates}.}

\item{alpha}{Type I error rate. Default is set at 0.05 to produce 95\%
confidence intervals.}

\item{odbc_connection}{an open odbc channel to CDMS database}

\item{cdms_host}{URL for CDMS website. The default is the Nez Perce Tribe's
data repository \url{https://npt-cdms.nptfisheries.org}.}
}
\value{
coordinated assessments NOSA DES table
}
\description{
Format NOSA data generated from \code{get_NOSAestimates}
  for the coordinated assessments natural origin spawner abundance DES
  tables. The function requires an odbc connection from behind the NPT
  firewall to download CAX metadata tables from the CDMS database.
  Additionally, if a dataframe is not supplied a connection to
  \url{https://npt-cdms.nptfisheries.org} using
  \code{cdmsR::cdmsLogin()} is also required.
}
\examples{
cdmsR::cdmsLogin('your_username', 'your_password')
df <- get_NOSAestimates()
con <- RODBC::odbcConnect(dsn = 'data source', uid = 'your_username', pwd = 'your_password')
format_NOSAdes(df, odbc_connection = con)
}
