% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_JuvenileDES.R
\name{format_JuvenileDES}
\alias{format_JuvenileDES}
\title{Format juvenile out-migrant DES tables.}
\usage{
format_JuvenileDES(
  df = NULL,
  des = c("detail", "smoltEq"),
  alpha = 0.05,
  odbc_connection,
  cdms_url = "https://cdms.nptfisheries.org"
)
}
\arguments{
\item{df}{dataframe output from \code{get_JuvenileEstimates}.}

\item{des}{designates the coordinated assessments DES table you are
constructing; \code{c("detail", "smoltEq")}.}
}
\value{
a dataframe with fish indicators and metrics, associated coordinated
  assessments population metadata, and location and project specific
  information
}
\description{
Format juvenile abundance and survival data generated from
  \code{get_JuvenileEstimates} for the coordinated assessments juvenile
  outmigrant and juvenile outmigrant detail DES tables. The function requires
  an odbc connection from behind the NPT firewall to download CAX metadata
  tables from the CDMS database.  Additionally, if a dataframe is not
  supplied a connection to \url{https://cdms.nptfisheries.org} using
  \code{cdmsR::cdmsLogin()} is also required.
}
\examples{
cdmsR::cdmsLogin('your_username', 'your_password')
df <- get_JuvenileEstimates()
con <- RODBC::odbcConnect(dsn = 'CDMS_PROD', uid = 'guest', pwd = 'guest')
format_JuvenileDES(df, des = 'smoltEq')
}
