% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CarcassAge.R
\name{get_carcassAge}
\alias{get_carcassAge}
\title{Carcass Age Data}
\usage{
get_carcassAge(
  carcass_data,
  age_data = NULL,
  rmis_codes = NULL,
  pit_codes = NULL,
  vie_marks = NULL
)
}
\arguments{
\item{carcass_data}{cleaned CDMS carcass dataset}

\item{age_data}{cleaned CDMS age summary dataset}

\item{rmis_codes}{RMIS release coded wire codes}

\item{pit_codes}{PTAGIS tagging detail query}

\item{vie_marks}{VIE marks applied by release year and potential return year}
}
\value{

}
\description{
Add scale, age fin, CWT, PIT and VIE age data to carcasses. Scale and age fin data is stored in the CDMS age summary dataset and connected to carcasses using stream name, survey date and sample number. CWT, PIT and VIE age data is supplied by the user and connect through code numbers (CWT and PIT) and stream name and return year for VIE marks.
}
\examples{
car_dat <- getDatasetView(datastoreID = 79)
car_dat <- clean_carcassData(car_dat)
age_dat <- getDatasetView(datastoreID = 80)
get_carcassAge(car_dat, age_dat)
}
\author{
Ryan N. Kinzer
}
