% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FinsTrapping.R
\name{get_FinsTrapping}
\alias{get_FinsTrapping}
\title{Summarize Fins Trapping Data}
\usage{
get_FinsTrapping(facility, odbc_connection)
}
\arguments{
\item{facility}{a vector of valid \url{https://www.finsnet.org/} hatchery
facilities}

\item{odbc_connection}{an open odbc channel to CDMS database}
}
\value{
a data.frame containing raw or summarized
  \url{https://www.finsnet.org/} trapping module information
}
\description{
Query and summarize \url{https://www.finsnet.org/} trapping
  module data. Currently the function requires an odbc connection from behind
  the NPT firewall to download trapping data that has been copied from
  \url{https://www.finsnet.org/}. Data is updated monthly or as needed.  In
  the future the function will retrieve data from a
  \url{https://www.finsnet.org/} API.
}
\examples{
con <- RODBC::odbcConnect(dsn = 'data source',
 uid = 'your_username', pwd = 'your_password')
get_FinsTrapping('NPT Hatchery', odbc_connection = con)
}
