% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_JuvenileDES.R
\name{get_JuvenileDES}
\alias{get_JuvenileDES}
\title{Download juvenile abundance data from CDMS and format for Coordinated
Assessments Juvenile Detail DES. The function requires a CDMS login to
first gather data.}
\usage{
get_JuvenileDES(
  output = c("detail", "smoltEq"),
  alpha = 0.05,
  cdms_url = "https://cdms.nptfisheries.org"
)
}
\arguments{
\item{alpha}{Type I error rate. Default is set at 0.05 to produce 95\%
confidence intervals.}

\item{dsId_abundance}{CDMS datasetId for RST abundance dataset.}

\item{dsID_survival}{CDMS datasetId for juvenile survival dataset.}

\item{cdms_host}{URL for CDMS website.}
}
\value{

}
\description{
Download juvenile abundance data from CDMS and format for Coordinated
Assessments Juvenile Detail DES. The function requires a CDMS login to
first gather data.
}
