% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_WeirData.R
\name{get_WeirData}
\alias{get_WeirData}
\title{Get Fins Trapping Data}
\usage{
get_WeirData(facility, odbc_connection)
}
\arguments{
\item{facility}{a vector of valid \url{https://www.finsnet.org/} hatchery
facilities}

\item{odbc_connection}{an open odbc channel to CDMS database}
}
\value{
a data.frame containing raw \url{https://www.finsnet.org/} trapping module information
}
\description{
Query \url{https://www.finsnet.org/} trapping
  module data. This function is set up to query directly from FINS API or from a static table on NPT server. Currently the function requires an odbc connection from behind
  the NPT firewall to download trapping data that has been copied from
  \url{https://www.finsnet.org/}. Data is updated monthly or as needed.  In
  the future the function will retrieve data from a
  \url{https://www.finsnet.org/} API.
}
\examples{
con <- RODBC::odbcConnect(dsn = 'Fins', uid = 'guest', pwd = 'guest')
qry <- "SELECT distinct(Facility) FROM FINS_all_trapping"
f <- RODBC::sqlQuery(con, qry) \%>\% mutate_all(as.character) \%>\% pull(Facility)
npt_f <- f[grepl('NPT', f)]
get_WeirData(npt_f, odbc_connection = con)
}
