% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_strata.R
\name{gg_strata}
\alias{gg_strata}
\title{Plotting function to help visualize strata for RST trap efficiencies}
\usage{
gg_strata(daily_df, strata_df)
}
\arguments{
\item{daily_df}{Summarized P4 RST data with daily C/M/R tallies and strata.}

\item{strata_df}{Daily data further summarized by strata.}

\item{strata_dates}{Vector of dates (YYYY-MM-DD)input by user to establish
strata.  If NULL, function will automatically apply one-week strata.}
}
\description{
Function inside 'sum_RSTstrata()', Takes summarized RST data, both
daily and by strata, and plots captures/marks/recaptures to help assess
accuracy of strata.
}
\examples{
p4_raw <- get_P4Data(EventSite = 'IMNTRP', MigrationYear = 2020)
p4_clean <- clean_P4Data(p4_raw)
strata <- sum_RSTstrata(p4_clean, 'Spring', NULL)
}
\author{
Tyler T. Stright
}
