% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guid.R
\name{guid}
\alias{guid}
\title{Generate GUID}
\usage{
guid(uppercase = FALSE)
}
\arguments{
\item{uppercase}{}
}
\value{
A 36 character long version 4.0 globally unique identifier.
}
\description{
Originally shared by Christopher Bare on R-bloggers;
  \url{https://www.r-bloggers.com/generate-uuids-in-r/}. The function was
  adapted to provide a 36 character GUID.
}
\examples{
guid(uppercase = TRUE)
guid(uppercase = FALSE)
}
