% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markrecap_flatfile.R
\name{markrecap_flatfile}
\alias{markrecap_flatfile}
\title{Title}
\usage{
markrecap_flatfile(carcass.data, species = "Chinook", run = "Spring/Summer")
}
\arguments{
\item{run}{}
}
\value{

}
\description{
Title
}
