% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_abundance_phat.R
\name{proportion_abundance_phat}
\alias{proportion_abundance_phat}
\title{Title}
\usage{
proportion_abundance_phat(N, SE.N, phat, SE.phat, alpha = 0.05)
}
\arguments{
\item{alpha}{}
}
\description{
Title
}

