% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_P4Data.R
\name{qa_P4Data}
\alias{qa_P4Data}
\title{qa_P4Data}
\usage{
qa_P4Data(data)
}
\arguments{
\item{data}{cleaned P4 data from get_RSTData() %>% clean_RSTData() and filtered for a single RST}

\item{srr_code}{vector of species run rear codes, can be a single code or multiple, if left NULL the function will return list of table for each srr_code}
}
\description{
Provides basic summaries and figures of P4 data to assist in data validation and the QA/QC process.
}
\examples{
p4_raw <- get_P4Data(EventSite = 'IMNTRP', MigrationYear = 2020)
p4_clean <- clean_P4Data(p4_raw)
qa_dat <- qa_P4Data(p4_clean)
names(qa_dat)
qa_dat[[1]]

}
\author{
Ryan N. Kinzer
}
