% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redd_expansion.R
\name{redd_expansion}
\alias{redd_expansion}
\title{Title}
\usage{
redd_expansion(R, n_f, n_s, n_p, n_fp, CI.type = "delta", alpha = 0.05,
  iterations = 1000, print = TRUE)
}
\arguments{
\item{print}{}
}
\description{
Title
}

