% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redd_sums.R
\name{redd_sums}
\alias{redd_sums}
\title{Title}
\usage{
redd_sums(
  redd.data,
  species = "Chinook",
  run = "Spring/Summer",
  probabilistic = FALSE
)
}
\arguments{
\item{probabilistic}{}
}
\value{

}
\description{
Title
}
