% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_assign.R
\name{strata_assign}
\alias{strata_assign}
\title{strata_assign}
\usage{
strata_assign(x, h_n)
}
\arguments{
\item{x}{vector of consecutive days for splitting into equal lengths}

\item{h_n}{number of days contained within each strata}
}
\value{

}
\description{
Create strata groups that have an equal number of days, except for
the last group.  The last group is truncated to be the same length as the
dataset.
}
\examples{
strata_assign(1:75,h_n = 7)
}
\author{
Ryan N. Kinzer
}
