% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_collapse.R
\name{strata_collapse}
\alias{strata_collapse}
\title{strata_collapse}
\usage{
strata_collapse(x, recaps)
}
\arguments{
\item{x}{vector of recaps}

\item{recaps}{number of designated recaps}
}
\value{

}
\description{
Collapses an initial strata scheme (e.g., 7 days) into strata containing at least the designated number of recaps. Seven day strata may become 14 days if recaps in the first strata are less than designated number.
}
\examples{
x <- c(1,1,1,4,1,1,2,2,2,1,6,7)
strata_collapse(x, 7)
}
\author{
Ryan N. Kinzer
}
