% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_groups.R
\name{sum_groups}
\alias{sum_groups}
\title{Sum Groups by Count Field}
\usage{
sum_groups(.data, .summary_var, .cnt_var, ...)
}
\arguments{
\item{.data}{dataset}

\item{...}{grouping variables}

\item{.sum_var}{variable to sum}
}
\value{

}
\description{
Sum a count field for records in a dataset within defined groups.
}
\author{
Ryan N. Kinzer
}
