% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_JUVests.R
\name{sum_JUVests}
\alias{sum_JUVests}
\title{Summarize Juvenile Abundance and Survival Metrics}
\usage{
sum_JUVests(
  alpha = c("0.05", "0.10"),
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{alpha}{Type I error rate. Default is set at 0.05 to produce 95\%
confidence intervals.}

\item{cdms_host}{URL for CDMS website. The default is the Nez Perce Tribe's
data repository \url{https://cdms.nptfisheries.org}.}
}
\value{
\code{get_JuvenileEstimates()} returns a dataframe with all lifestage
  abundance, survival and smolt equivalent (smoltEq) estimates stored in
  \url{https://cdms.nptfisheries.org} for each juvenile screw trap, seining
  and hatchery release group.
}
\description{
Download juvenile abundance and survival estimates from NPT's
  data repository, \url{https://cdms.nptfisheries.org}. Requires a valid
  username and password to access data.
}
\examples{
cdmsR::cdmsLogin('your_username', 'your_password')
sum_JUVests()
}
