% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_Lamprey.R
\name{sum_Lamprey}
\alias{sum_Lamprey}
\title{sum_Lamprey:}
\usage{
sum_Lamprey(data)
}
\arguments{
\item{data}{cleaned P4 data queried for a single calendar year.}
}
\value{

}
\description{
Summarize P4 lamprey catch for BPA deliverable.
}
\examples{
data_raw <- get_P4Data(SRRcode = 'A0W', CalendarYear = 2021)
data_clean <- cuyem::clean_p4Data(data_raw)
lamprey_sum <- sum_Lamprey(data_clean
}
\author{
Tyler T. Stright
}
