% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_NOSAests.R
\name{sum_NOSAests}
\alias{sum_NOSAests}
\title{Estimate NOSA DES}
\usage{
sum_NOSAests(redd_data, carcass_data, mr_ests, split_age = FALSE)
}
\arguments{
\item{redd_data}{clean redd data}

\item{carcass_data}{clean carcass data with best age appended}

\item{mr_ests}{mark-recapture data}
}
\value{

}
\description{
Estimate NOSA DES
}
\examples{
sum_NOSAests()
}
\author{
Ryan N. Kinzer
}
