% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_Redds.R
\name{sum_Redds}
\alias{sum_Redds}
\title{Sum Redd Counts}
\usage{
sum_Redds(.redd_data, ...)
}
\arguments{
\item{...}{grouping variables}

\item{redd_data}{cleaned redd data}
}
\value{

}
\description{
Sum Redd Counts
}
\examples{
redd_df <- cdmsR::getDatasetView(datastoreID = 78)
redd_dat <- clean_ReddData(redd_df)
sum_Redds(redd_dat, SurveyYear, StreamName, TransectName)
}
\author{
Ryan N. Kinzer
}
