% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_SGSests.R
\name{sum_SGSests}
\alias{sum_SGSests}
\title{Summarize SGS Data}
\usage{
sum_SGSests(redd_data = NULL, carcass_data = NULL, ...)
}
\arguments{
\item{redd_data}{cleaned redd data}

\item{carcass_data}{cleaned carcass data}

\item{...}{variables to group the data}
}
\value{

}
\description{
Summarize and estimate performance measures for SGS data.
}
\examples{
sum_SGSests(clean_redd, clean_car)
}
\author{
Ryan N. Kinzer, Tyler T. Stright
}
