% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_WaterTempData.R
\name{sum_WaterTempData}
\alias{sum_WaterTempData}
\title{sum_WaterTempData:}
\usage{
sum_WaterTempData(data)
}
\arguments{
\item{data}{CDMS water temperature data from get_WaterTempData() %>% clean_WaterTempData()}
}
\value{
dataframe containing summarized water temperature metrics
}
\description{
summarizes cleaned water temperature data
}
\examples{
wt_raw <- get_WaterTempData(date_begin = '2009-07-01', date_end = '2009-11-30', locationID = 1605)
wt_clean <- clean_WaterTempData(wt_raw)
wt_sum <- sum_WaterTempData(wt_clean)
}
