% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_WaterTemps.R
\name{sum_WaterTemps}
\alias{sum_WaterTemps}
\title{Summarize Water Temperature Data}
\usage{
sum_WaterTemps(data)
}
\arguments{
\item{data}{CDMS water temperature data from get_WaterTemps() %>% clean_WaterTemps()}
}
\value{
dataframe containing summarized water temperature metrics
}
\description{
summarizes cleaned water temperature data
}
\examples{
wt_raw <- getWaterTemps(date_begin = '2009-07-01', date_end = '2009-11-30', locationID = 1605)
wt_clean <- clean_WaterTemps(wt_raw)
wt_sum <- sum_WaterTemps(wt_clean)
}
