% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstream_abundance.R
\name{upstream_abundance}
\alias{upstream_abundance}
\title{Estimate adults abundance upstream of a weir using mark-recapture data.}
\usage{
upstream_abundance(C, M, R, groups = NULL, alpha = 0.05,
  CI.type = c("normal", "boot"), iter = 1000, print = TRUE)
}
\arguments{
\item{C}{a vector of captures at the second capture occasion}

\item{M}{a vector of marks applied at the first capture occasion}

\item{R}{a vector of recaptures as the second capture occasion}

\item{groups}{optional character vector of strata group names}

\item{alpha}{Type I error rate.  Default is set at 0.05 to produce 95\%
confidence intervals.}

\item{CI.type}{Confidence intervals can either be reported as asymptotic normal or bootstrapped intervals.}

\item{iter}{The number of bootstrap iterations to be run.  Default is set at
1000.}

\item{print}{Logical value. Default value is TRUE. Should results be printed to the console?}
}
\description{
\code{upstream_abundance} generates stratified adult abundance
  estimates from mark-recapture data collected at Chinook salmon and steelhead
  pickett weirs or during Chinook salmon spawning ground surveys. The function
  using the adjusted Lincoln-Peterson estimator to account for small sample
  bias.

#@section Warning:
}
\examples{

C <- c(100,200,300)
M <- c(50,100,150)
R <-  c(10,20,30)
s_name <- c('small', 'medium', 'large')
tmp <- upstream_abundanc(C,M,R,groups = s_name, CI.type = 'normal')

}
\author{
Ryan N. Kinzer
}

