% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid.R
\name{uuid}
\alias{uuid}
\title{Generate UUID}
\usage{
uuid(uppercase = FALSE)
}
\arguments{
\item{uppercase}{}
}
\value{
A 40 character long version 4.0 universal unique identifier.
}
\description{
Originally shared by Christopher Bare on R-bloggers;
  \url{https://www.r-bloggers.com/generate-uuids-in-r/}.
}
\examples{
uuid(uppercase = TRUE)
uuid(uppercase = FALSE)
}
