% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_addition.R
\name{vector_addition}
\alias{vector_addition}
\title{Title}
\usage{
vector_addition(
  fx = "multiply",
  type = "both_random",
  X,
  SE.X,
  Y,
  SE.Y = NULL,
  alpha = 0.05
)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
Title
}
