% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_FCRRData.R
\name{clean_FCRRData}
\alias{clean_FCRRData}
\title{Clean Fall Chinook Run Reconstruction (FCRR) Data}
\usage{
clean_FCRRData(data)
}
\arguments{
\item{data}{raw CDMS FCRR dataset from \code{cdmsR::getDatasetView(datastore = 100)} or from CDMS dataset export}
}
\description{
Processes raw FCRR data into a more manageable format.
}
\examples{
fcrr_dat <- getDatasetView(datastore = 100)
fcrr_clean <- clean_FCRRData(fcrr_dat)
}
\author{
Tyler T. Stright
}
