% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_P4data.R
\name{clean_P4Data}
\alias{clean_P4Data}
\title{clean_P4Data:}
\usage{
clean_P4Data(data)
}
\arguments{
\item{data}{raw CDMS P4 data from get_P4data()}
}
\description{
Processes raw P4 data and adds important fields for summaries
}
\examples{
p4_raw <- get_P4Data()
p4_clean <- clean_P4Data(p4_raw)
}
\author{
Tyler T. Stright
}
