% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_JuvenileDES.R
\name{format_JuvenileDES}
\alias{format_JuvenileDES}
\title{Format juvenile out-migrant DES tables.}
\usage{
format_JuvenileDES(
  df = NULL,
  alpha = c(0.05, 0.1),
  odbc_connection,
  cdms_url = "https://cdms.nptfisheries.org"
)
}
\arguments{
\item{df}{dataframe output from \code{get_JuvenileEstimates}.}

\item{alpha}{Type I error rate. Default is set at 0.05 to produce 95\%
confidence intervals.}

\item{odbc_connection}{an open odbc channel to CDMS database}

\item{cdms_host}{URL for CDMS website. The default is the Nez Perce Tribe's
data repository \url{https://cdms.nptfisheries.org}.}
}
\value{
a list with coordinated assessments juvenile outmigrant and juvenile outmigrant detail DES tables
}
\description{
Format juvenile abundance and survival data generated from
  \code{get_JuvenileEstimates} for the coordinated assessments juvenile
  outmigrant and juvenile outmigrant detail DES tables. The function requires
  an odbc connection from behind the NPT firewall to download CAX metadata
  tables from the CDMS database.  Additionally, if a dataframe is not
  supplied a connection to \url{https://cdms.nptfisheries.org} using
  \code{cdmsR::cdmsLogin()} is also required.
}
\examples{
cdmsR::cdmsLogin('your_username', 'your_password')
df <- get_JuvenileEstimates()
con <- RODBC::odbcConnect(dsn = 'data source', uid = 'your_username', pwd = 'your_password')
format_JuvenileDES(df, odbc_connection = con)
}
