% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SGSests.R
\name{get_SGSests}
\alias{get_SGSests}
\title{Summarize SGS Data}
\usage{
get_SGSests(redd_data = NULL, carcass_data = NULL, group)
}
\arguments{
\item{redd_data}{cleaned redd data}

\item{carcass_data}{cleaned carcass data with best age appended}

\item{group}{variables to group the data}

\item{species}{Chinook or steelhead}

\item{run}{Spring/summer, fall, summer}
}
\value{

}
\description{
Summarize and estimate performance measures for SGS data.
}
\examples{
get_SGSests(clean_redd, clean_car)
}
\author{
Ryan N. Kinzer
}
