% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_strata.R
\name{gg_strata}
\alias{gg_strata}
\title{gg_strata:}
\usage{
gg_strata(daily_df, strata_df, species = c("Chinook", "Steelhead"))
}
\arguments{
\item{daily_df}{Summarized P4 RST data with daily C/M/R tallies and strata.
Produced by sum_RSTstrata().}

\item{strata_df}{Daily data further summarized by strata. Produced by sum_RSTstrata().}

\item{strata_dates}{Vector of dates (YYYY-MM-DD) input by user to establish
strata.  If NULL, function will automatically apply 7-day strata.}
}
\description{
Function inside 'sum_RSTstrata()', Takes summarized RST data, both
daily and by strata, and plots captures/marks/recaptures to help assess
accuracy of strata. Not intended for use outside of sum_RSTstrata().
}
\author{
Tyler T. Stright
}
